change WaitCmd behaviour to wait for inhibited mask bits to be released so that command complete. Use WaitCmd throught the code (perhaps rework WaitCmd to DoCmd and have it perform the Send and wait).
git-svn-id: https://svn.aros.org/svn/aros/trunk@46898 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/devs/sdcard/sdcard_intern.h
@@ -231,8 +231,8 @@ void FNAME_SDCBUS(BusTask)(struct sdcard_Bus *);
 BOOL FNAME_SDC(RegisterVolume)(struct sdcard_Bus *);
 
 ULONG FNAME_SDCBUS(WaitUnitStatus)(ULONG, struct sdcard_Unit *);
-int FNAME_SDCBUS(SDSCChangeFrequency)(struct sdcard_Unit *);
-int FNAME_SDCBUS(MMCChangeFrequency)(struct sdcard_Unit *);
+ULONG FNAME_SDCBUS(SDSCChangeFrequency)(struct sdcard_Unit *);
+ULONG FNAME_SDCBUS(MMCChangeFrequency)(struct sdcard_Unit *);
 
 BOOL FNAME_SDC(HandleIO)(struct IORequest *io);
